Move include cookies function
